Modarb-AI/requirements.txt
opencv-python-headless
datasets
transformers 
accelerate 
evaluate  
bitsandbytes 
accelerate 
einops
Pillow
torch
torchvision